+++ src/routes/form/form.test.js
@@ -2,5 +2,5 @@ import app from '../../app';
 import { testRequest } from '../../helpers';
 
 describe('Form', () => {
-  it('should submit a form', () => testRequest(app, 'search/post'));
+  it('should submit a form', () => testRequest(app, 'form/post'));
 });